zebra_cli/rest_client/pyproject.toml
[tool.poetry]
name = "io-t-connector-local-rest-ap-is-client"
version = "1.0.0"
description = "A client library for accessing IoT Connector Local REST APIs"
authors = []
readme = "README.md"
packages = [
    {include = "io_t_connector_local_rest_ap_is_client"},
]
include = ["CHANGELOG.md", "io_t_connector_local_rest_ap_is_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.23.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
